refactor(exec): move result structs into `run.ts` and `output.ts`
