+++ .eslintrc.json
@@ -10,7 +10,8 @@
   "rules": {
     "camelcase": "off",
     "i18n-text/no-en": "off",
-    "import/no-namespace": "off"
+    "import/no-namespace": "off",
+    "no-shadow": "off"
   },
   "env": {
     "es6": true,
