+++ packages/exec/lib/output.d.ts
@@ -1,4 +1,15 @@
-import { OutputResult } from "./result";
+import { RunResult } from "./run";
+/** A command run and output get result */
+export declare class OutputResult extends RunResult {
+    /** The log output */
+    output: string;
+    /**
+     * Constructs a new command run and output get result
+     * @param code the status code
+     * @param output the log output
+     */
+    constructor(code: number, output: string);
+}
 /**
  * Runs a command and gets the output
  * @param command a command to run
