+++ packages/exec/lib/run.d.ts
@@ -1,4 +1,18 @@
-import { RunResult } from "./result";
+/** A command run result */
+export declare class RunResult {
+    /** The status code */
+    code: number;
+    /**
+     * Constructs a new command run result
+     * @param code the status code
+     */
+    constructor(code: number);
+    /**
+     * Checks if the status is ok (status code is `0`)
+     * @returns `true` if the status is ok
+     */
+    isOk(): boolean;
+}
 /**
  * Runs a command
  * @param command a command to run
